# docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateRoomBody:
    properties:
      name:
        example: Pazarlama Ekibi
        type: string
    required:
    - name
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: invalid request parameters
        type: string
    type: object
...
  handler.InviteMemberBody:
    properties:
      email:
        example: teammate@example.com
        type: string
      role:
        description: '"writer" or "reader"'
        example: writer
        type: string
    required:
    - email
    - role
    type: object
  handler.RoomMemberResponse:
    properties:
      email:
        type: string
      nickname:
        type: string
      role:
        type: string
      user_id:
        type: integer
    type: object
  handler.RoomResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      members:
        items:
          $ref: '#/definitions/handler.RoomMemberResponse'
        type: array
      name:
        type: string
      owner_id:
        type: integer
      role:
        type: string
    type: object
...
  model.Room:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      owner_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.RoomMember:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role:
        enum:
        - writer
        - reader
        type: string
      room_id:
        type: integer
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@talkforge.local
    name: TalkForge Support
  description: Production-ready authentication and role verification API for TalkForge.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TalkForge API
  version: "1.0"
paths:
...
  /api/v1/rooms:
    get:
      description: Returns every room the authenticated user is a member of.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.RoomResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my rooms
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Creates a room that talks can be shared inside; the creator becomes
        its first writer member.
      parameters:
      - description: Create Room Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateRoomBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a shared room
      tags:
      - Rooms
  /api/v1/rooms/{id}:
    get:
      description: Returns a room's details and member list. Requires membership.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RoomResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get room details
      tags:
      - Rooms
  /api/v1/rooms/{id}/members:
    post:
      consumes:
      - application/json
      description: Adds (or updates the role of) a user in a room by email. Requires
        the requester to be a writer member.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Invite Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.InviteMemberBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RoomMember'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite a member into a room
      tags:
      - Rooms
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"